iio: adc: sun4i-gpadc-iio: add support for H5 thermal sensor
This patch adds support for the H5 ths sensor.

The H5 supports interrupts. The interrupt is configured to update
the sensor values every second. The sensor is except the scale/offset
values and an additional sensor for the gpu the same. A temperatur higer
than 70°C needs an other formula. Since the operating temperatures maximum
is 70°C this case was neglegted.

Signed-off-by: Philipp Rossak <embed3d@gmail.com>